requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
aiofiles==24.1.0
websockets>=13.0
pydantic-settings==2.6.0